feat: enhance cart routes

+++ services/repository/cart_repository.go
@@ -1,6 +1,7 @@
 package repository
 
 import (
+	"errors"
 	"microservices/db"
 	"microservices/models"
 )
@@ -19,8 +20,21 @@ func UpdateCartItem(cart *models.Cart) error {
 	return db.DB.Save(cart).Error
 }
 
-func RemoveCartItem(cartID uint) error {
-	return db.DB.Delete(&models.Cart{}, cartID).Error
+func RemoveCartItem(cartID uint, quantity int, userId uint) error {
+	var cart models.Cart
+	err := db.DB.Where("id = ?", cartID).First(&cart).Error
+	if err != nil {
+		return err
+	}
+	if cart.UserID != userId {
+		return errors.New("cart item not found")
+	}
+
+	if quantity >= cart.Quantity {
+		return db.DB.Delete(&models.Cart{}, cartID).Error
+	}
+
+	return db.DB.Model(&cart).Update("quantity", cart.Quantity-quantity).Error
 }
 
 func ClearCart(userID uint) error {
